requirements.txt
Flask
python-dotenv
openai
azure-cognitiveservices-speech
Flask-CORS 